Extract style helpers to separate module
