+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2016-11-07 19:04:38 +0300 using RuboCop version 0.45.0.
+# on 2016-11-08 12:47:19 +0300 using RuboCop version 0.45.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -23,7 +23,7 @@ Metrics/AbcSize:
 # Offense count: 1
 # Configuration parameters: CountComments.
 Metrics/ClassLength:
-  Max: 739
+  Max: 704
 
 # Offense count: 3
 Metrics/CyclomaticComplexity:
